Modify Post and Topic class to resolve conflict

+++ src/main/java/com/mangalaxy/mango/repository/TopicRepository.java
@@ -1,7 +1,8 @@
 package com.mangalaxy.mango.repository;
 
-import com.mangalaxy.mango.model.entity.Topic;
+import com.mangalaxy.mango.domain.entity.Topic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TopicRepository extends JpaRepository<Topic, Long> {
+public interface TopicRepository extends JpaRepository<Topic, Short> {
+
 }
